add service for governance and other features
